Split Code and Object to different files
TBD: support assparser to parse lyrics without k commands
